Doc: Remove stale documentation in "qproperty.cpp"
"src/corelib/kernel/qproperty.cpp" documents assignment operators over
`QPropertyBinding` for `QProperty` and `QPropertyAlias`.

Those operators were removed in 638df6138e7acecc2d3b6aad0771e2d81355f2d2
and are thus stale.

Similarly, an assignment operator for `QPropertyAlias<T>` over a `T&&`
is documented albeit it was removed in
927647cd032cd2e43bae3184b879586849ffee50.

Due an incorrect implementation of QDoc, which has now been removed,
with regards to identifying copy/move assignment operators, and the more
superficial and incorrect way in which QDoc binds those kind of elements
to the clang AST, QDoc was unable to report the lack of a binding
declaration for those elements that were removed, producing incorrect
documentation instead.

As of the latest changes to QDoc this will correctly generate a warning.

To avoid the addition of new warnings and remove stale and incorrect
documentation, the document blocks were removed from the file.

Change-Id: I30e20f26d4c44e4589abf40546d8ebaa4e869638
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>