wasm: fix accessibility crash on null parent
The code was a assuming that an a11y interface always
has a parent, which is not the case for the root interface.
Also factor out to a getWindow() function, to prepare
for re-use later on.

Change-Id: I52a841bf94c712deb9603e8b9c2c878820c5f117
Reviewed-by: Mikołaj Boc <Mikolaj.Boc@qt.io>
(cherry picked from commit 34c36821578f39e9c7f99e1dd0e7887a52418b9d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>